Update crypto script

+++ scripts/deploy_cryptomining-tool.sh
@@ -5,14 +5,19 @@ ENVIRONMENT=${ENVIRONMENT:-dev}
 DEVOPS_CONFIG=${DEVOPS_CONFIG:-devops-config}
 VALUES=${DEVOPS_CONFIG}/environments/${ENVIRONMENT}/cryptomining_values.yaml
 
-helm delete --purge cryptomining-detector
-result=$(echo $?)
-if [[ result == 1 ]]; then
-  echo "Could not purge existing chart"
-  exit 1
+helm ls | grep cryptomining-detector
+if [[ $? == 0 ]]; then
+  helm delete --purge cryptomining-detector
+  result=$(echo $?)
+  if [[ result == 1 ]]; then
+    echo "Could not purge existing chart"
+    exit 1
+  else
+    echo "Purged existing chart"
+    sleep 60
+  fi
 else
-  echo "Purged existing chart"
-  sleep 60
+  echo "char for cryptomining detector doesn't exist"
 fi
 
 kubectl get deployments -n "${NAMESPACE}" | grep cryptomining-detector
